+++ tests/imodels-clients-tests/src/integration/authoring/LockOperations.test.ts
@@ -7,31 +7,24 @@ import { expect } from "chai";
 import {
   ConflictingLock,
   ConflictingLocksError,
-  GetLockListParams,
   IModelsClient,
   IModelsClientOptions,
-  Lock,
   LockLevel,
-  LockLevelFilter,
   LocksError,
   UpdateLockParams,
 } from "@itwin/imodels-client-authoring";
 import {
   AuthorizationCallback,
   IModelsErrorCode,
-  toArray,
 } from "@itwin/imodels-client-management";
 import {
   IModelMetadata,
-  ReusableIModelMetadata,
-  ReusableTestIModelProvider,
   TestAuthorizationProvider,
   TestIModelCreator,
   TestIModelFileProvider,
   TestIModelGroup,
   TestIModelGroupFactory,
   TestUtilTypes,
-  assertCollection,
   assertError,
   assertLock,
 } from "@itwin/imodels-client-test-utils";
@@ -44,7 +37,6 @@ describe("[Authoring] LockOperations", () => {
 
   let testIModelFileProvider: TestIModelFileProvider;
   let testIModelGroup: TestIModelGroup;
-  let testIModelForRead: ReusableIModelMetadata;
   let testIModelForWrite: IModelMetadata;
   let testIModelForWriteBriefcaseIds: number[] = [];
 
@@ -68,11 +60,6 @@ describe("[Authoring] LockOperations", () => {
       testSuiteName: "AuthoringLockOperations",
     });
 
-    const reusableTestIModelProvider = container.get(
-      ReusableTestIModelProvider
-    );
-    testIModelForRead = await reusableTestIModelProvider.getOrCreate();
-
     const testIModelCreator = container.get(TestIModelCreator);
     testIModelForWrite = await testIModelCreator.createEmptyAndUploadChangesets(
       testIModelGroup.getPrefixedUniqueIModelName("Test iModel for write")
@@ -93,98 +80,6 @@ describe("[Authoring] LockOperations", () => {
   after(async () => {
     await testIModelGroup.cleanupIModels();
   });
-
-  it("should return all items when querying collection", async () => {
-    // Arrange
-    const getLockListParams: GetLockListParams = {
-      authorization,
-      iModelId: testIModelForRead.id,
-    };
-
-    // Act
-    const locks = iModelsClient.locks.getList(getLockListParams);
-
-    // Assert
-    await assertCollection({
-      asyncIterable: locks,
-      isEntityCountCorrect: (count) => count === 1,
-    });
-  });
-
-  it("should return correct values when querying collection with briefcaseId filter", async () => {
-    // Arrange
-    const getLockListParams: GetLockListParams = {
-      authorization,
-      iModelId: testIModelForRead.id,
-      urlParams: {
-        briefcaseId: testIModelForRead.briefcases[0].id,
-      },
-    };
-
-    // Act
-    const locks = iModelsClient.locks.getList(getLockListParams);
-
-    // Assert
-    const lockArray = await toArray(locks);
-    const lock = lockArray[0];
-    assertLock({
-      actualLock: lock,
-      expectedLock: testIModelForRead.lock,
-    });
-  });
-
-  it("should return empty collection when querying locks for non-existent briefcase", async () => {
-    // Arrange
-    const getLockListParams: GetLockListParams = {
-      authorization,
-      iModelId: testIModelForRead.id,
-      urlParams: {
-        briefcaseId: 500,
-      },
-    };
-
-    // Act
-    const locks = iModelsClient.locks.getList(getLockListParams);
-
-    // Assert
-    const lockArray = await toArray(locks);
-    expect(lockArray.length).to.equal(0);
-  });
-
-  ([LockLevel.Shared, LockLevel.Exclusive] as LockLevelFilter[]).forEach(
-    (lockLevel) => {
-      it(`should return correct values when querying collection with lockLevel filter (${lockLevel})`, async () => {
-        // Arrange
-        const expectedLock: Lock = {
-          briefcaseId: testIModelForRead.lock.briefcaseId,
-          lockedObjects: testIModelForRead.lock.lockedObjects.filter(
-            (lockedObject) => lockedObject.lockLevel === lockLevel
-          ),
-        };
-        expect(expectedLock.lockedObjects.length).to.be.greaterThan(0);
-
-        const getLockListParams: GetLockListParams = {
-          authorization,
-          iModelId: testIModelForRead.id,
-          urlParams: {
-            lockLevel,
-          },
-        };
-
-        // Act
-        const locks = iModelsClient.locks.getList(getLockListParams);
-
-        // Assert
-        const lockArray = await toArray(locks);
-        expect(lockArray.length).to.be.equal(1);
-        const actualLock = lockArray[0];
-        assertLock({
-          actualLock,
-          expectedLock,
-        });
-      });
-    }
-  );
 
   it("should acquire new locks", async () => {
     // Arrange