+++ itwin-platform-access/imodels-access-backend/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @itwin/imodels-access-backend
 
-This log was last generated on Thu, 15 Jan 2026 01:26:07 GMT and should not be manually modified.
+This log was last generated on Thu, 23 Apr 2026 08:41:44 GMT and should not be manually modified.
+
+## 6.1.1
+Thu, 23 Apr 2026 08:41:44 GMT
+
+### Patches
+
+- Use new release Locks in chunks api to release all Locks
 
 ## 6.1.0
 Thu, 15 Jan 2026 01:26:07 GMT